Factor tt::is_maplike into its own file
