+++ tests/Utilities/MakeWithRandomValues.hpp
@@ -18,6 +18,7 @@
 #include "Utilities/Requires.hpp"
 #include "Utilities/TypeTraits.hpp"
 #include "Utilities/TypeTraits/GetFundamentalType.hpp"
+#include "Utilities/TypeTraits/IsMaplike.hpp"
 
 namespace TestHelpers_detail {
 /// \cond HIDDEN_SYMBOLS
